# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RedForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/scan:
    post:
      operationId: triggerScan
      tags: [projects]
      summary: Trigger a scan for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
...
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Scan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /scans/{id}/logs:
    get:
      operationId: getScanLogs
      tags: [scans]
      summary: Get scan logs (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE stream of scan logs
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Scan:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        findingsCount:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - status
        - findingsCount
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - createdAt
...
    ScanDetail:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED]
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        findingsCount:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        findings:
          type: array
          items:
            $ref: "#/components/schemas/FindingWithProject"
        logs:
          type: array
          items:
            $ref: "#/components/schemas/ScanLog"
      required:
        - id
        - projectId
        - projectName
        - status
        - findingsCount
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - createdAt
        - findings
        - logs
    ScanLog:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        level:
          type: string
          enum: [INFO, WARN, ERROR, DEBUG]
        message:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - level
        - message
        - createdAt
...
    FindingWithProject:
      type: object
      properties:
        id:
          type: string
        scanId:
          type: string
        projectId:
          type: string
        projectName:
          type: string
        title:
          type: string
        description:
          type: string
        endpoint:
          type: string
        severity:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW, INFO]
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, FIXED, WONT_FIX, FALSE_POSITIVE]
        cvss:
          type: number
          nullable: true
        cwe:
          type: string
          nullable: true
        owasp:
          type: string
          nullable: true
        pocCode:
          type: string
          nullable: true
        fixPatch:
          type: string
          nullable: true
        fixExplanation:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - scanId
        - projectId
        - projectName
        - title
        - description
        - endpoint
        - severity
        - status
        - createdAt
        - updatedAt
...